+++ src/ewokscore/methodtask.py
@@ -1,5 +1,8 @@
-from .task import Task
+from typing import Mapping
+
 from ewoksutils.import_utils import import_method
+
+from .task import Task
 
 METHOD_ARGUMENT = "_method"
 
@@ -8,6 +11,9 @@ class MethodExecutorTask(
     Task, input_names=[METHOD_ARGUMENT], output_names=["return_value"]
 ):
     METHOD_ARGUMENT = METHOD_ARGUMENT
+
+    def _get_task_identifier(self, inputs: Mapping) -> str:
+        return inputs.get(self.METHOD_ARGUMENT, self.class_registry_name())
 
     def run(self):
         kwargs = self.get_named_input_values()